pyproject.toml

[project]
name = "gcpdiag"
version = "0.1.0"
description = "GcpDiag"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "authlib==1.6.1",
    "beautifulsoup4==4.13.4",
    "blessings==1.7",
    "boltons==25.0.0",
    "bs4==0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cryptography==45.0.5",
    "cyclopts==3.22.3",
    "dataclasses==0.6",
    "diskcache==5.6.3",
    "dnspython==2.7.0",
    "docstring-parser==0.17.0",
    "docutils==0.22",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "fastmcp==2.10.6",
    "frozenlist==1.7.0",
    "google-adk>=1.10.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp==1.12.2",
    "mdurl==0.1.2",
    "multidict==6.6.3",
    "oauthlib==3.3.1",
    "openapi-pydantic==0.5.1",
    "packaging>=23.2",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydantic[email]==2.11.7",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pyperclip==1.9.0",
    "pytest>=8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "ratelimit==2.2.1",
    "referencing==0.36.2",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "six>=1.15.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sse-starlette==3.0.2",
    "starlette==0.47.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "vertexai>=1.43.0",
    "yarl==1.20.1",
]

[tool.pytest.ini_options]
pythonpath = ["gcpdiag"]

[tool.uv]
package = false

[tool.setuptools]
packages = ["gcpdiag"]
